+++ src/main/java/org/openlmis/referencedata/web/FacilityController.java
@@ -36,7 +36,7 @@ import org.openlmis.referencedata.domain.SupervisoryNode;
 import org.openlmis.referencedata.domain.SupplyLine;
 import org.openlmis.referencedata.domain.SupportedProgram;
 import org.openlmis.referencedata.dto.ApprovedProductDto;
-import org.openlmis.referencedata.dto.BasicFacilityDto;
+import org.openlmis.referencedata.dto.MinimalFacilityDto;
 import org.openlmis.referencedata.dto.FacilityDto;
 import org.openlmis.referencedata.dto.SupportedProgramDto;
 import org.openlmis.referencedata.exception.NotFoundException;
@@ -127,10 +127,10 @@ public class FacilityController extends BaseController {
   @RequestMapping(value = "/facilities", method = RequestMethod.GET)
   @ResponseStatus(HttpStatus.OK)
   @ResponseBody
-  public List<BasicFacilityDto> getAllFacilities() {
+  public List<FacilityDto> getAllFacilities() {
     rightService.checkAdminRight(RightName.FACILITIES_MANAGE_RIGHT);
 
-    return toBasicDto(facilityRepository.findAll());
+    return toDto(facilityRepository.findAll());
   }
 
 
@@ -325,12 +325,12 @@ public class FacilityController extends BaseController {
   @RequestMapping(value = "/facilities/search", method = RequestMethod.POST)
   @ResponseStatus(HttpStatus.OK)
   @ResponseBody
-  public Page<BasicFacilityDto> searchFacilities(@RequestBody Map<String, Object> queryParams,
-                                            Pageable pageable) {
+  public Page<MinimalFacilityDto> searchFacilities(@RequestBody Map<String, Object> queryParams,
+                                                   Pageable pageable) {
     rightService.checkAdminRight(RightName.FACILITIES_MANAGE_RIGHT);
 
     List<Facility> foundFacilities = facilityService.searchFacilities(queryParams);
-    List<BasicFacilityDto> facilityDtos = toBasicDto(foundFacilities);
+    List<MinimalFacilityDto> facilityDtos = toBasicDto(foundFacilities);
     return Pagination.getPage(facilityDtos, pageable);
   }
 
@@ -359,14 +359,14 @@ public class FacilityController extends BaseController {
     return productDtos;
   }
 
-  private BasicFacilityDto toBasicDto(Facility facility) {
-    BasicFacilityDto dto = new BasicFacilityDto();
-    facility.basicExport(dto);
+  private MinimalFacilityDto toBasicDto(Facility facility) {
+    MinimalFacilityDto dto = new MinimalFacilityDto();
+    facility.export(dto);
 
     return dto;
   }
 
-  private List<BasicFacilityDto> toBasicDto(Iterable<Facility> facilities) {
+  private List<MinimalFacilityDto> toBasicDto(Iterable<Facility> facilities) {
     return StreamSupport
         .stream(facilities.spliterator(), false)
         .map(this::toBasicDto)